Add Units and Temperature Resources
- Measurement, Temperature, and Unit models
- Temperature and Unit tables using Table Per Concrete Class (TPC)
- Unit resource endpoints with integration tests

+++ Odin.Hardware/ArduinoTMP36ButtonSerial/JsonDefinitions.cs
@@ -3,7 +3,7 @@ namespace Odin.Hardware.ArduinoTMP36ButtonSerial;
 public record ArduinoTMP36ReadingJson
 {
     public required int DeviceId { get; init; }
-    public required float DegreesCelsius { get; init; }
+    public required double DegreesCelsius { get; init; }
 }
 
 public record PostRequestJsonBody : ArduinoTMP36ReadingJson
